Model formula

Build a model formula for an MMRM.

brm_formula(
  data,
  intercept = TRUE,
  effect_baseline = TRUE,
  effect_group = TRUE,
  effect_time = TRUE,
  interaction_baseline = TRUE,
  interaction_group = TRUE,
  correlation = "unstructured"
)

Arguments

data

A classed data frame from brm_data().

intercept

TRUE to include an intercept, FALSE to omit.

effect_baseline

TRUE to include an additive effect for baseline response, FALSE to omit.

effect_group

TRUE to include an additive effects for treatment groups, FALSE to omit.

effect_time

TRUE to include a additive effect for discrete time points, FALSE to omit.

interaction_baseline

TRUE to include baseline-by-time interaction, FALSE to omit.

interaction_group

TRUE to include treatment-group-by-time interaction, FALSE to omit.

correlation

Character of length 1, name of the correlation structure. Only "unstructured" is currently supported.

Value

An object of class "brmsformula" returned from brms::brmsformula(). It contains the fixed effect parameterization, correlation structure, and residual variance structure.

Parameterization

The formula is not the only factor that determines the fixed effect parameterization. The ordering of the categorical variables in the data, as well as the contrast option in R, affect the construction of the model matrix. To see the model matrix that will ultimately be used in brm_model(), run brms::make_standata() and examine the X element of the returned list. See the examples below for a demonstration.

See also

Examples

set.seed(0)
data <- brm_data(
  data = brm_simulate_simple()$data,
  outcome = "response",
  role = "response",
  group = "group",
  time = "time",
  patient = "patient",
  level_control = "group_1",
  level_baseline = "time_1"
)
brm_formula(data)
#> response ~ time + group + group:time + unstr(time = time, gr = patient) 
#> sigma ~ 0 + time
brm_formula(data = data, intercept = FALSE, effect_baseline = FALSE)
#> response ~ 0 + time + group + group:time + unstr(time = time, gr = patient) 
#> sigma ~ 0 + time
formula <- brm_formula(
  data = data,
  intercept = FALSE,
  effect_baseline = FALSE,
  interaction_group = FALSE
)
formula
#> response ~ 0 + time + group + unstr(time = time, gr = patient) 
#> sigma ~ 0 + time
# Optional: set the contrast option, which determines the model matrix.
options(contrasts = c(unordered = "contr.SAS", ordered = "contr.poly"))
# See the fixed effect parameterization you get from the data:
head(brms::make_standata(formula = formula, data = data)$X)
#>   timetime_1 timetime_2 timetime_3 timetime_4 groupgroup_1
#> 1          1          0          0          0            1
#> 2          0          1          0          0            1
#> 3          0          0          1          0            1
#> 4          0          0          0          1            1
#> 5          1          0          0          0            1
#> 6          0          1          0          0            1
# Specify a different contrast method to use an alternative
# parameterization when fitting the model with brm_model():
options(
  contrasts = c(unordered = "contr.treatment", ordered = "contr.poly")
)
# different model matrix than before:
head(brms::make_standata(formula = formula, data = data)$X)
#>   timetime_1 timetime_2 timetime_3 timetime_4 groupgroup_2
#> 1          1          0          0          0            0
#> 2          0          1          0          0            0
#> 3          0          0          1          0            0
#> 4          0          0          0          1            0
#> 5          1          0          0          0            0
#> 6          0          1          0          0            0