MCMC draws from the marginal posterior of an MMRM

Get marginal posterior draws from a fitted MMRM.

brm_marginal_draws(
  model,
  data,
  use_subgroup = !is.null(attr(data, "brm_subgroup")),
  control = NULL,
  baseline = NULL
)

Arguments

model

Fitted brms model object from brm_model().

data

Classed tibble with preprocessed data from brm_data().

use_subgroup

Logical of length 1, whether to summarize the draws by each subgroup level. If TRUE, subgroup-specific marginals are given. Otherwise, the subgroup is marginalized out.

control

Deprecated. Set the control group level in brm_data().

baseline

Deprecated. Set the control group level in brm_data().

Value

A named list of tibbles of MCMC draws of the marginal posterior distribution of each treatment group and time point (or group-by-subgroup-by-time, if applicable). In each tibble, there is 1 row per posterior sample sand one column for each type of marginal distribution (i.e. each combination of treatment group and discrete time point. The specific tibbles in the returned list are described below:

response: on the scale of the response variable.

difference_time: change from baseline: the response at a particular time minus the response at baseline (reference_time). Only returned if the role argument of brm_data() was "response". (If role is "change", then response already represents change from baseline.)

difference_group: treatment effect: the the difference_time at each active group minus the difference_time at the control group (reference_group). If role is "change", then treatment group is instead the difference between response at each active group minus the response at the control group.

difference_subgroup: subgroup differences: the difference_group at each subgroup level minus the difference_group at the subgroup reference level (reference_subgroup).

Separation string

Post-processing in brm_marginal_draws() names each of the group-by-time marginal means with the delimiting character string from Sys.getenv("BRM_SEP", unset = "|"). Neither the column names nor element names of the group and time variables can contain this string. To set a custom string yourself, use Sys.setenv(BRM_SEP = "YOUR_CUSTOM_STRING").

See also

Examples

if (identical(Sys.getenv("BRM_EXAMPLES", unset = ""), "true")) {
set.seed(0L)
data <- brm_data(
  data = brm_simulate_simple()$data,
  outcome = "response",
  role = "response",
  group = "group",
  time = "time",
  patient = "patient",
  reference_group = "group_1",
  reference_time = "time_1"
)
formula <- brm_formula(
  data = data,
  baseline = FALSE,
  baseline_time = FALSE
)
tmp <- utils::capture.output(
  suppressMessages(
    suppressWarnings(
      model <- brm_model(
        data = data,
        formula = formula,
        chains = 1,
        iter = 100,
        refresh = 0
      )
    )
  )
)
brm_marginal_draws(model = model, data = data)
}