Append simulated categorical covariates

Simulate and append non-time-varying categorical covariates to an existing brm_data() dataset.

Usage

brm_simulate_categorical(data, names, levels, probabilities = NULL)

Arguments

data

Classed tibble as from brm_data() or brm_simulate_outline().

names

Character vector with the names of the new covariates to simulate and append. Names must all be unique and must not already be column names of data.

levels

Character vector of unique levels of the simulated categorical covariates.

probabilities

Either NULL or a numeric vector of length length(levels) with levels between 0 and 1 where all elements sum to 1. If NULL, then all levels are equally likely to be drawn. If not NULL, then probabilities is a vector of sampling probabilities corresponding to each respective level of levels.

Value

A classed tibble, like from brm_data() or brm_simulate_outline(), but with new categorical covariate columns and with the names of the new covariates appended to the brm_covariates attribute. Each new categorical covariate column is a character vector, not the factor type in base R.

Details

Each covariate is a new column of the dataset with one independent random categorical draw for each patient, using a fixed set of levels (via base::sample() with replace = TRUE). All covariates simulated this way are independent of everything else in the data, including other covariates (to the extent that the random number generators in R work as intended).

See also

Examples

data <- brm_simulate_outline()
brm_simulate_categorical(
  data = data,
  names = c("site", "region"),
  levels = c("area1", "area2")
)
#> # A tibble: 800 × 7
#>    response missing group   time   patient     site  region
#>       <dbl> <lgl>   <chr>   <chr>  <chr>       <chr> <chr> 
#>  1       NA FALSE   group_1 time_1 patient_001 area2 area2 
#>  2       NA FALSE   group_1 time_2 patient_001 area2 area2 
#>  3       NA FALSE   group_1 time_3 patient_001 area2 area2 
#>  4       NA FALSE   group_1 time_4 patient_001 area2 area2 
#>  5       NA FALSE   group_1 time_1 patient_002 area1 area2 
#>  6       NA FALSE   group_1 time_2 patient_002 area1 area2 
#>  7       NA TRUE    group_1 time_3 patient_002 area1 area2 
#>  8       NA TRUE    group_1 time_4 patient_002 area1 area2 
#>  9       NA FALSE   group_1 time_1 patient_003 area2 area2 
#> 10       NA FALSE   group_1 time_2 patient_003 area2 area2 
#> # ℹ 790 more rows
brm_simulate_categorical(
  data = data,
  names = c("site", "region"),
  levels = c("area1", "area2"),
  probabilities = c(0.1, 0.9)
)
#> # A tibble: 800 × 7
#>    response missing group   time   patient     site  region
#>       <dbl> <lgl>   <chr>   <chr>  <chr>       <chr> <chr> 
#>  1       NA FALSE   group_1 time_1 patient_001 area2 area2 
#>  2       NA FALSE   group_1 time_2 patient_001 area2 area2 
#>  3       NA FALSE   group_1 time_3 patient_001 area2 area2 
#>  4       NA FALSE   group_1 time_4 patient_001 area2 area2 
#>  5       NA FALSE   group_1 time_1 patient_002 area2 area2 
#>  6       NA FALSE   group_1 time_2 patient_002 area2 area2 
#>  7       NA TRUE    group_1 time_3 patient_002 area2 area2 
#>  8       NA TRUE    group_1 time_4 patient_002 area2 area2 
#>  9       NA FALSE   group_1 time_1 patient_003 area2 area2 
#> 10       NA FALSE   group_1 time_2 patient_003 area2 area2 
#> # ℹ 790 more rows